\name{aal90}
\docType{data}
\alias{aal90}
\title{Coordinates for data from the AAl90 atlas}
\description{
  This is a list of spatial coordinates for the AAL90 atlas, along with
  indices for the major lobes of the brain.
}
\usage{aal90}
\format{List, 17 elements}
\references{
  Tzourio-Mazoyer N., Landeau B., Papathanassiou D., Crivello F., Etard O.,
  Delcroix N., Mazoyer B., Joliot M. (2002) \emph{Automated anatomical labeling
  of activations in SPM using a macroscopic anatomical parcellation of the
  MNI MRI single-subject brain}. NeuroImage, 15(1):273-289.
}
\keyword{datasets}
