% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{args_as_list}
\alias{args_as_list}
\title{Convert arguments into a single list}
\usage{
args_as_list(...)
}
\arguments{
\item{...}{Graphs or a list of graphs}
}
\value{
A list object
}
\description{
\code{args_as_list} converts its arguments into a single list. If the
argument is already a single list, it returns the same object. If multiple
objects are given, it wraps them in a list.
}
\keyword{internal}
