% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainGraph_mediate.R
\name{bg_to_mediate}
\alias{bg_to_mediate}
\title{Convert brainGraph results to "mediate" object}
\usage{
bg_to_mediate(x, region = NULL)
}
\arguments{
\item{x}{Object output from \code{\link{brainGraph_mediate}}}

\item{region}{Character string specifying which region's results to
summarize; only relevant if \code{level='vertex'} (default: \code{NULL})}
}
\value{
An object of class \code{mediate}
}
\description{
\code{\link{bg_to_mediate}} converts the results of
\code{\link{brainGraph_mediate}} into an object of class
\code{\link[mediation]{mediate}}. In \code{brainGraph}, it is only used for
the \code{\link[mediation]{summary.mediate}} method, but you can similarly
use its output for the \code{\link[mediation]{plot.mediate}} method.
}
\seealso{
\code{\link[mediation]{mediate}}
}
