% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{brainGraph}
\alias{brainGraph}
\alias{brainGraph-options}
\alias{brainGraph-package}
\title{Default options for brainGraph}
\description{
brainGraph is a package for performing \emph{graph theory analysis} of brain
MRI data.
}
\section{Package options}{


brainGraph uses the following \code{\link{options}} to configure behavior:
\itemize{
  \item \code{bg.subject_id}: character string specifying the name your
    project/study uses as a subject identifier. All imported data (e.g.,
    covariates tables) \emph{MUST} have a column matching this. One possible
    alternative is \code{'participant_id'}, recommended by BIDS. Default:
    \code{'Study.ID'}
  \item \code{bg.group}: character string specifying the name your
    project/study uses as a group identifier. All imported data (e.g.,
    covariates tables) \emph{MUST} have a column matching this. One possible
    alternative is \code{'group'}, recommended by BIDS. Default:
    \code{'Group'}
  \item \code{bg.session}: character string specifying the name your
    project/study uses as a \dQuote{time} or session identifier, in the case
    of longitudinal studies. All imported data (e.g., covariates tables)
    \emph{MUST} have a column matching this. One possible alternative is
    \code{'session_id'}, recommended by BIDS. Default: \code{'Time'}
  \item \code{bg.progress}: logical indicating whether to show progress bars
    for functions that provide the option. Default: \code{TRUE}
}
}

