\name{brainsuite}
\docType{data}
\alias{brainsuite}
\title{Coordinates for data from the BrainSuite software}
\description{
  This is a list of spatial coordinates for the BrainSuite software, along with
  indices for the major lobes of the brain.
}
\usage{brainsuite}
\format{List, 17 elements}
\references{
  Shattuck D.W. and Leahy R.M. (2002) \emph{BrainSuite: an automated cortical
  surface identification tool}. Medical Image Analysis, 8(2):129-142.
}
\keyword{datasets}
