% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_weights}
\alias{check_weights}
\alias{check_degree}
\alias{check_strength}
\title{Check for vertex or edge attributes}
\usage{
check_weights(g, weights)

check_degree(g)

check_strength(g)
}
\value{
\code{check_weights} - If \code{weights=NULL} and the graph has a
  \code{'weight'} attribute, then \code{NULL} will be returned. If
  \code{weights=NA}, then \code{NA} is returned.
}
\description{
\code{check_weights} is a helper function for dealing with edge weights that
get passed to different \code{igraph} functions.

\code{check_degree} is a helper function to check if \code{degree} is a
vertex attribute of the input graph. Returns a numeric vector of the degree
values.

\code{check_strength} is a helper function to check if \code{strength} is a
vertex attribute of the input graph. Returns a numeric vector of the strength
values.
}
\keyword{internal}
