% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_degree}
\alias{check_degree}
\title{Check for presence of a degree attribute}
\usage{
check_degree(g)
}
\description{
Helper function to check if \code{degree} is a vertex attribute of the input
graph. Returns a numeric vector of the degree values.
}
\keyword{internal}
