% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_strength}
\alias{check_strength}
\title{Check for presence of a strength attribute}
\usage{
check_strength(g)
}
\description{
Helper function to check if \code{strength} is a vertex attribute of the
input graph. Returns a numeric vector of the strength values.
}
\keyword{internal}
