% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_weights}
\alias{check_weights}
\title{Check for edge weights}
\usage{
check_weights(g, weights)
}
\description{
\code{check_weights} is a helper function for dealing with edge weights that
get passed to different \code{igraph} functions.
}
\details{
If \code{weights=NULL} and the graph has a \code{'weight'} attribute, that
will be returned. If \code{weights=NA}, then that is returned.
}
\keyword{internal}
