% Generated by roxygen2 (4.0.1): do not edit by hand
\name{color.edges}
\alias{color.edges}
\title{Color the edges of a graph based on community membership.}
\usage{
color.edges(g, comm)
}
\arguments{
\item{g}{The graph to get its edges colored}

\item{comm}{The community object returned from community.measures}
}
\value{
A character vector of colors for each edge in the graph
}
\description{
This function takes the community membership of a given graph, and assigns
to the edges a specific color (the same as the vertex membership colors).
Currently only works for up to 9 communities (colored: red, green, blue,
yellow, magenta, orangered, lightgreen, lightblue, lightyellow). Edges that
connect two different communities are colored gray.
}

