% Generated by roxygen2 (4.0.1): do not edit by hand
\name{color.vertices}
\alias{color.vertices}
\title{Give graph vertices a color attribute.}
\usage{
color.vertices(comm, atlas = NULL)
}
\arguments{
\item{comm}{A \link{communities} object}

\item{atlas}{A character string}
}
\value{
A character vector of length \code{vcount(g)}
}
\description{
This function takes a \link{communities} object (or a character string
indicating the brain atlas used (e.g. for the major lobes of the brain), and
creates a character vector of colors for each vertex.
}

