% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GraphColors}
\alias{GraphColors}
\alias{set_vertex_color}
\alias{set_edge_color}
\title{Color graph vertices and edges}
\usage{
set_vertex_color(g, name, memb)

set_edge_color(g, name, memb)
}
\arguments{
\item{g}{An \code{igraph} graph object}

\item{name}{Character string of the name of the vertex attribute to add}

\item{memb}{An integer vector representing membership of e.g. a community}
}
\value{
The same graph with additional vertex or edge attribute
}
\description{
\code{set_vertex_color} takes an integer vector representing membership of
some grouping (e.g., a community or connected component) and creates a
character vector of colors for each grouping. Isolated vertices will be
colored \emph{gray}.

\code{set_edge_color} assigns a color to each edge (the same as the vertex
membership colors). Edges that connect vertices of two different groups are
colored gray.
}
\keyword{internal}
