% Generated by roxygen2 (4.0.0): do not edit by hand
\name{community.measures}
\alias{community.measures}
\title{Community/module measures.}
\usage{
community.measures(g)
}
\arguments{
\item{g}{the adjacency graph}
}
\value{
A list with the following components:
\item{community}{The \link{communities} object for the present graph.}
\item{vcolors}{A list of the vertex colors for the 8 largest communities.}
}
\description{
This function splits a graph into communities, and gives vertices for each
community a separate color.
}

