\name{coords.brainnet.dk68.split2}
\docType{data}
\alias{coords.brainnet.dk68.split2}
\title{Coordinates for Freesurfer data}
\description{
  This is a data frame of spatial coordinates for the DK atlas, when each label
  has been split into two. The coordinates are in MNI space and suitable for use
  with the BrainNet Viewer.
}
\usage{coords.brainnet.dk68.split2}
\format{Data frame, 3 columns (x, y, and z) and 68 observations}
\references{
  Desikan R.S., Segonne F., Fischl B., et al. (2006) \emph{An automated
  labeling system for subdividing the human cerebral cortex on MRI scans into
  gyral based regions of interest}. NeuroImage, 31:968-980.
}
\keyword{datasets}
