\name{coords.brainnet.dkt62}
\docType{data}
\alias{coords.brainnet.dkt62}
\title{Coordinates for Freesurfer data}
\description{
  This is a data frame of spatial coordinates for the DKT atlas. The
  coordinates are in MNI space and suitable for use with the BrainNet Viewer.
}
\usage{coords.brainnet.dkt62}
\format{Data frame, 3 columns (x, y, and z) and 68 observations}
\references{
  Klein A. and Tourville J. (2012) \emph{101 labeled brain images and a
  consistent human cortical labeling protocol}. Front Neurosci,
  doi:10.3389/fnins.2012.00171
}
\keyword{datasets}
