\name{coords.dkt}
\docType{data}
\alias{coords.dkt}
\title{Coordinates for Freesurfer data}
\description{
  This is a data frame of spatial coordinates for the DKT atlas.
}
\usage{coords.dkt}
\format{Data frame, 3 columns (x, y, and z) and 62 observations}
\references{
  Klein A. and Tourville J. (2012) \emph{101 labeled brain images and a
  consistent human cortical labeling protocol}. Front Neurosci,
  doi:10.3389/fnins.2012.00171
}
\keyword{datasets}
