\name{coords.ic2}
\docType{data}
\alias{coords.ic2}
\title{Coordinates for Freesurfer data}
\description{
  This is a data frame of spatial coordinates for labels split based on their
  icosahedral face each vertex maps to. Based on \emph{ic2.tri} in Freesurfer.
}
\usage{coords.ic2}
\format{Data frame, 3 columns (x, y, and z) and 297 observations}
\keyword{datasets}
