\name{coords.split2}
\docType{data}
\alias{coords.split2}
\title{Coordinates for Freesurfer data}
\description{
  This is a data frame of spatial coordinates when each label of the DK atlas
  (from Freesurfer) was split into two labels.
}
\usage{coords.split2}
\format{Data frame, 3 columns (x, y, and z) and 136 observations}
\keyword{datasets}
