% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr.matrix}
\alias{corr.matrix}
\title{Calculate correlation matrix and threshold}
\usage{
corr.matrix(dat, density, thresh = NULL, exclusions = NULL, ...)
}
\arguments{
\item{dat}{Data table of the data to correlate}

\item{density}{Numeric indicating the resultant network density; keeps the
top \emph{X}\% of correlations}

\item{thresh}{Numeric; absolute correlation value to threshold by (default:
\code{NULL})}

\item{exclusions}{Numeric vector of indices (columns) to exclude (default:
\code{NULL})}

\item{...}{Other arguments, passed to \code{\link[Hmisc]{rcorr}}}
}
\value{
A list with the following components:
\item{R}{Numeric matrix of correlation coefficients.}
\item{P}{Numeric matrix of p-values.}
\item{r.thresh}{Binary matrix indicating correlations that are above a
  certain threshold.}
\item{threshold}{Numeric; the threshold value used.}
}
\description{
\code{corr.matrix} calculates the correlation between all column pairs of a
given data frame, and thresholds the resultant correlation matrix based on a
given density (e.g., \code{0.1} if you want to keep only the 10\% strongest
correlations.
}
\details{
If you wish to exclude regions from your analysis, you can give the indices
of their columns with the \code{exclusions} argument.

By default, the Pearson correlation coefficients are calculated, but can
return Spearman by passing an additional argument.
}
\examples{
\dontrun{
corrs <- lapply(groups, function(x) lapply(densities, function(y)
  corr.matrix(resids.all[x], density=y)))
}
}
\seealso{
\code{\link[Hmisc]{rcorr}}

Other Volumetric functions: \code{\link{brainGraph_init}},
  \code{\link{check.resid}}, \code{\link{get.resid}}
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
