% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/count_homologous.R
\name{count_homologous}
\alias{count_homologous}
\title{Count number of edges between homologous regions of a brain graph}
\usage{
count_homologous(g)
}
\arguments{
\item{g}{An igraph graph object}
}
\value{
A named vector of the edge ID's connecting homologous regions
}
\description{
This function will count the number of edges between homologous regions in a
brain graph (e.g. between L and R superior frontal).
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}

