% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_all_attr}
\alias{delete_all_attr}
\title{Delete all attributes of a graph}
\usage{
delete_all_attr(g, keep.names = FALSE)
}
\arguments{
\item{g}{An \code{igraph} graph object}

\item{keep.names}{Logical indicating whether to keep the \code{name} vertex
attribute (default: \code{FALSE})}
}
\value{
An \code{igraph} graph object
}
\description{
Deletes all graph-, vertex-, and edge-level attributes of an \code{igraph}
graph objects.
}
\seealso{
\code{\link[igraph]{delete_graph_attr},
  \link[igraph]{delete_vertex_attr}, \link[igraph]{delete_edge_attr}}
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
