% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mats.R
\name{dir2files}
\alias{dir2files}
\title{Return a vector of filenames based on a directory name or options list}
\usage{
dir2files(x)
}
\arguments{
\item{x}{Either a character string specifying a directory or a list of
arguments to be passed to \code{\link{list.files}}}
}
\description{
Return a vector of filenames based on a directory name or options list
}
\keyword{internal}
