% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resid.R
\name{get_lm_vars}
\alias{get_lm_vars}
\title{Get some variables for LM}
\usage{
get_lm_vars(covars, exclude, ...)
}
\arguments{
\item{covars}{A \code{data.table} of the covariates of interest}

\item{exclude}{Character vector of covariates to exclude (default:
\code{NULL})}

\item{...}{Arguments passed to \code{\link{brainGraph_GLM_design}} (optional)}
}
\value{
A list containing:
  \item{X}{The design matrix}
  \item{lev}{The leverage}
  \item{n}{The number of observations}
  \item{p}{The number of parameters}
}
\description{
Get some variables for LM
}
\keyword{internal}
