% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata}
\alias{get_metadata}
\title{Add metadata to a list-like object}
\usage{
get_metadata(object)
}
\arguments{
\item{object}{A list-like object}
}
\value{
The same object with version, system, and date information added
}
\description{
Adds metadata to a list-like object.
}
\details{
If the object is a graph, graph-level attributes will be added. The
elements added are:
\itemize{
  \item{version}{A list with R, brainGraph, and igraph versions}
  \item{sys}{Character vector of system information}
  \item{date}{The date and time of creation}
}
}
\keyword{internal}
