% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_random_graphs.R
\name{get_rand_attrs}
\alias{get_rand_attrs}
\title{Convenience function to get attributes for lists of random graphs}
\usage{
get_rand_attrs(bg.list, level)
}
\arguments{
\item{bg.list}{A \code{brainGraphList} object, created in
\code{\link{analysis_random_graphs}}}
}
\value{
A \code{data.table}
}
\description{
Convenience function to get attributes for lists of random graphs
}
\keyword{internal}
