% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainGraph_GLM.R
\name{glm_fit_helper}
\alias{glm_fit_helper}
\title{Helper function for GLM fitting}
\usage{
glm_fit_helper(DT, X, con.type, con.mat, alt, measure, mykey)
}
\description{
Helper function for GLM fitting
}
\keyword{internal}
