% Generated by roxygen2 (4.0.1): do not edit by hand
\name{global.eff}
\alias{global.eff}
\title{Calculate the global efficiency of a graph.}
\usage{
global.eff(g)
}
\arguments{
\item{g}{The adjacency graph on which to calculate efficiency}
}
\value{
A number; the global efficiency of the graph.
}
\description{
This function calculates the global efficiency of a graph.
}
\references{
Latora V., Marchiori M. (2001) \emph{Efficient behavior of
small-world networks}. Phys Rev Lett, 87.19:198701.
}

