% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_contract_brain.R
\name{graph.contract.brain}
\alias{graph.contract.brain}
\title{Contract graph vertices based on brain lobe and hemisphere}
\usage{
graph.contract.brain(g)
}
\arguments{
\item{g}{The graph to contract}
}
\value{
A new graph
}
\description{
This function creates a new graph after merging multiple vertices based on
brain lobe and hemisphere membership. The new vertex size is equal to the
number of vertices in each lobe. The x- and y- coordinates of the new
vertices are equal to the mean of the lobe vertices of the original graph.
The new edge weight is equal to the number of inter-lobular connections of
the original graph.
}
\seealso{
\code{\link[igraph]{contract.vertices}}
}

