% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_attr_dt.R
\name{graph_attr_dt}
\alias{graph_attr_dt}
\title{Create a data table with graph global measures}
\usage{
graph_attr_dt(g.list, group = NULL)
}
\arguments{
\item{g.list}{A list of igraph graph objects}

\item{group}{A character string indicating group membership (default:NULL)}
}
\value{
A data table with several columns (equal to the number of graph
attributes) and row number equal to the number of graphs in the input list
}
\description{
This is just a helper function that takes a list of graphs and creates a data
table of global measures for each graph, ordered by graph density.
}
\seealso{
\code{\link[igraph]{graph_attr}, \link[igraph]{graph_attr_names}}
}
