% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hist_thickness.R
\name{hist_thickness}
\alias{hist_thickness}
\title{Plot group histograms of cortical thickness for a given brain region}
\usage{
hist_thickness(dat, cur.region)
}
\arguments{
\item{dat}{A data table of cortical thicknesses; needs columns for 'Group',
'region', and 'thickness'}

\item{cur.region}{A character string or integer of the brain region to plot;
if integer, the region is chosen from the data table based on the index}
}
\value{
A ggplot object for plotting the histograms
}
\description{
This function takes a "tidied" dataset of cortical thickness and plots a
histogram for 2 groups, including a dashed vertical line indicating the group
means.
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
\seealso{
\code{\link[ggplot2]{geom_histogram}, \link[ggplot2]{geom_vline}}
}

