\name{hoa112}
\docType{data}
\alias{hoa112}
\title{Coordinates for data from Harvard-Oxford atlas}
\description{
  This is a list of spatial coordinates for the Harvard-Oxford atlas, along with
  indices for the major lobes of the brain.
}
\usage{hoa112}
\format{List, 17 elements}
\references{
    Makris N., Goldstein J.M., Kennedy D. et al. (2006) \emph{Decreased volume
    of left and total anterior insular lobule in schizophrenia}. Schizophr Res,
    83(2-3):155-171.
}
\keyword{datasets}
