% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{is.brainGraph}
\alias{is.brainGraph}
\title{Determine whether x is a brainGraph object}
\usage{
is.brainGraph(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Determine whether x is a brainGraph object
}
\keyword{internal}
