% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_binary}
\alias{is_binary}
\title{Check if a matrix is binary}
\usage{
is_binary(mat)
}
\value{
Logical of length 1
}
\description{
Check if a matrix is binary
}
\seealso{
Other Matrix functions: \code{\link{apply_thresholds}},
  \code{\link{cor.diff.test}}, \code{\link{create_mats}},
  \code{\link{symmetrize_mats}}
}
\concept{Matrix functions}
\keyword{internal}
