% Generated by roxygen2 (4.0.1): do not edit by hand
\name{local.eff}
\alias{local.eff}
\title{Calculate the local efficiency of a graph.}
\usage{
local.eff(g)
}
\arguments{
\item{g}{the adjacency graph on which to calculate efficiency}
}
\value{
A vector of the local efficiencies for each node of the graph.
}
\description{
This function calculates the local efficiency of each node of a graph.
}
\references{
Latora V., Marchiori M. (2001) \emph{Efficient behavior of
small-world networks}. Phys Rev Lett, 87.19:198701.
}

