\name{lpba40}
\docType{data}
\alias{lpba40}
\title{Coordinates for brain MRIdata}
\description{
  This is a list of spatial coordinates for the LPBA40 atlas, along with
  indices for the major lobes of the brain. The coordinates were obtained
  from some colleagues.
}
\usage{lpba40}
\format{List, 17 elements}
\references{
  Shattuck DW, Mirza M, Adisetiyo V, Hojatkashani C, Salamon G, Narr KL,
  Poldrack RA, Bilder RM, Toga AW (2007) \emph{Construction of a 3D probabilistic
  atlas of human cortical structures}. NeuroImage,
  doi:10.1016/j.neuroimage.2007.09.031
}
\keyword{datasets}
