% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{make_brainGraph.bg_mediate}
\alias{make_brainGraph.bg_mediate}
\title{Create a graph with mediation-specific attributes}
\usage{
\method{make_brainGraph}{bg_mediate}(x, atlas = x$atlas,
  type = "observed", level = "contrast", set.attrs = FALSE,
  modality = NULL, weighting = NULL, threshold = NULL, ...)
}
\arguments{
\item{x}{A \code{bg_mediate} object}

\item{atlas}{Character string specifying the brain atlas}

\item{type}{Character string indicating the type of graphs. Default:
\code{observed}}

\item{level}{Character string indicating whether the graphs are subject-,
group-, or contrast-specific. Default: \code{'subject'}}

\item{set.attrs}{Logical indicating whether to assign all graph-, vertex-,
and edge-level attributes (via \code{\link{set_brainGraph_attr}}). Default:
\code{TRUE}}

\item{modality}{Character string indicating imaging modality (e.g. 'dti').
Default: \code{NULL}}

\item{weighting}{Character string indicating how the edges are weighted
(e.g., 'fa', 'pearson', etc.). Default: \code{NULL}}

\item{threshold}{Numeric indicating the threshold used when sparsifying the
connectivity matrix (if any). Default: \code{NULL}}

\item{...}{Other arguments passed to \code{\link{make_brainGraph}}}
}
\value{
A \code{brainGraph_mediate} graph object with attributes:
  \item{Graph}{\emph{mediator}, \emph{treat}, \emph{outcome}, \emph{nobs}}
  \item{Vertex}{\emph{b?.acme, p?.acme}, \emph{b?.ade, p?.ade},
    \emph{b?.prop, p?.prop}, \emph{b.tot, p.tot}}
}
\description{
This function only creates a graph for \emph{vertex}-level analyses.
}
\seealso{
Other Graph creation functions: \code{\link{CreateGraphs}},
  \code{\link{brainGraphList}},
  \code{\link{make_brainGraphList.NBS}},
  \code{\link{make_brainGraphList.bg_GLM}},
  \code{\link{make_ego_brainGraph}}
}
\concept{Graph creation functions}
