% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{make_brainGraphList.bg_GLM}
\alias{make_brainGraphList.bg_GLM}
\alias{make_brainGraphList.mtpc}
\title{Create a graph list with GLM-specific attributes}
\usage{
\method{make_brainGraphList}{bg_GLM}(x, atlas, type = "observed",
  level = "contrast", set.attrs = FALSE, modality = NULL,
  weighting = NULL, threshold = NULL, gnames = x$con.name, ...)

\method{make_brainGraphList}{mtpc}(x, atlas, type = "observed",
  level = "contrast", set.attrs = FALSE, modality = NULL,
  weighting = NULL, threshold = NULL, gnames = x$con.name, ...)
}
\arguments{
\item{x}{A \code{bg_GLM} or \code{mtpc} object}

\item{atlas}{Character string specifying the brain atlas to use}

\item{type}{Character string indicating the type of graphs. Default:
\code{observed}}

\item{level}{Character string indicating whether the graphs are subject-,
group-, or contrast-specific. Default: \code{'subject'}}

\item{set.attrs}{Logical indicating whether to assign all graph-, vertex-,
and edge-level attributes (via \code{\link{set_brainGraph_attr}}). Default:
\code{TRUE}}

\item{modality}{Character string indicating imaging modality (e.g. 'dti').
Default: \code{NULL}}

\item{weighting}{Character string indicating how the edges are weighted
(e.g., 'fa', 'pearson', etc.). Default: \code{NULL}}

\item{threshold}{Numeric indicating the threshold used when "sparsifying" the
connectivity matrix (if any). Default: \code{NULL}}

\item{gnames}{Character vector of graph names (e.g., study IDs if
\code{level='subject'}). Default: \code{NULL}}

\item{...}{Other arguments passed to \code{\link{set_brainGraph_attr}}}
}
\value{
A list of \code{igraph} graph objects (length equal to the number of
  contrasts) with additional attributes:
  \item{Graph}{\emph{name} (contrast name), \emph{outcome} (the outcome
    variable), \emph{alpha} (the significance level); for MTPC:
    \emph{tau.mtpc}, \emph{S.mtpc}, \emph{S.crit}, \emph{A.crit}}
  \item{Vertex}{\emph{size2} (t-statistic), \emph{size} (the t-stat
    transformed for visualization purposes), \emph{p} (equal to \eqn{1-p}),
    \emph{p.fdr} (equal to \eqn{1-p_{FDR}}, the FDR-adjusted p-value),
    \emph{gamma} (the contrast of parameter estimaties, \emph{se} (the
    standard error of \emph{gamma}); \emph{A.mtpc}, \emph{sig} (binary
    indicating whether \code{A.mtpc > A.crit}) (for MTPC)}
}
\description{
Create a \code{brainGraphList} with attributes specific to the results of
\code{\link{brainGraph_GLM}} or \code{\link{mtpc}}. The \code{graphs} element
of the returned object will contain one graph for each contrast.
}
\note{
Only valid for \emph{vertex}-level analyses.
}
\seealso{
\code{\link{brainGraph_GLM}, \link{mtpc}}

Other Graph creation functions: \code{\link{CreateGraphs}},
  \code{\link{brainGraphList}},
  \code{\link{make_brainGraph.bg_mediate}},
  \code{\link{make_brainGraphList.NBS}},
  \code{\link{make_ego_brainGraph}}
}
\concept{Graph creation functions}
