% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{make_brainGraphList.NBS}
\alias{make_brainGraphList.NBS}
\title{Create a graph list for NBS results}
\usage{
\method{make_brainGraphList}{NBS}(x, atlas, type = "observed",
  level = "contrast", set.attrs = TRUE, modality = NULL,
  weighting = NULL, threshold = NULL, gnames = x$con.name,
  groups = NULL, mode = "undirected", weighted = TRUE,
  diag = FALSE, ...)
}
\arguments{
\item{x}{A \code{NBS} object}

\item{atlas}{Character string specifying the brain atlas}

\item{type}{Character string indicating the type of graphs. Default:
\code{observed}}

\item{level}{Character string indicating whether the graphs are subject-,
group-, or contrast-specific. Default: \code{'subject'}}

\item{set.attrs}{Logical indicating whether to assign all graph-, vertex-,
and edge-level attributes (via \code{\link{set_brainGraph_attr}}). Default:
\code{TRUE}}

\item{modality}{Character string indicating imaging modality (e.g. 'dti').
Default: \code{NULL}}

\item{weighting}{Character string indicating how the edges are weighted
(e.g., 'fa', 'pearson', etc.). Default: \code{NULL}}

\item{threshold}{Numeric indicating the threshold used when sparsifying the
connectivity matrix (if any). Default: \code{NULL}}

\item{gnames}{Character vector of graph names (e.g., study IDs if
\code{level='subject'}). Default: \code{NULL}}

\item{groups}{Character (or factor) vector of group names (default:
\code{NULL})}

\item{mode}{Character string defining how the matrix should be interpreted.
Default: \code{'undirected'}}

\item{weighted}{Logical specifying whether to create a weighted network}

\item{diag}{Logical indicating whether to include the diagonal of the
connectivity matrix. Default: \code{FALSE}}

\item{...}{Other arguments passed to \code{\link{set_brainGraph_attr}}}
}
\value{
A list of \code{igraph} graph objects (length equal to the number of
  contrasts) with additional attributes:
  \item{Graph}{\emph{name} (contrast name)}
  \item{Vertex}{\emph{comp} (integer vector indicating connected component
    membership), \emph{p.nbs} (P-value for each component)}
  \item{Edge}{\emph{stat} (the test statistic for each connection), \emph{p}
    (the P-value)}
}
\description{
Create a graph list for NBS results
}
\seealso{
Other Graph creation functions: \code{\link{CreateGraphs}},
  \code{\link{brainGraphList}},
  \code{\link{make_brainGraph.bg_mediate}},
  \code{\link{make_brainGraphList.bg_GLM}},
  \code{\link{make_ego_brainGraph}}
}
\concept{Graph creation functions}
