% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{make_mediate_brainGraph}
\alias{make_mediate_brainGraph}
\title{Create a graph with mediation-specific attributes}
\usage{
make_mediate_brainGraph(res.med, atlas, ...)
}
\arguments{
\item{res.med}{List object output by \code{\link{brainGraph_mediate}}}

\item{atlas}{Character string specifying the brain atlas to use}

\item{...}{Other arguments passed to \code{\link{make_brainGraph}}}
}
\value{
A \code{brainGraph_mediate} graph object with attributes:
  \item{Graph}{\emph{mediator}, \emph{treat}, \emph{outcome}, \emph{nobs}}
  \item{Vertex}{\emph{b?.acme, p?.acme}, \emph{b?.ade, p?.ade},
    \emph{b?.prop, p?.prop}, \emph{b.tot, p.tot}}
}
\description{
This function only creates a graph for \emph{vertex}-level analyses.
}
\seealso{
Other Graph creation functions: \code{\link{make_brainGraph}},
  \code{\link{make_ego_brainGraph}},
  \code{\link{make_empty_brainGraph}},
  \code{\link{make_glm_brainGraph}},
  \code{\link{make_nbs_brainGraph}}
}
