% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix2list}
\alias{matrix2list}
\title{Convert a matrix to a list of rows}
\usage{
matrix2list(mat)
}
\arguments{
\item{mat}{Numeric matrix in which each row is a single contrast vector}
}
\value{
A list with length equal to the number of rows of \code{C}
}
\description{
Makes working with different contrast types (i.e., t or F) a little simpler.
}
\keyword{internal}
