\name{mni152}
\docType{data}
\alias{mni152}
\title{Nifti data for the MNI 152 brain MRI template}
\description{
  This object is a nifti object representing the MNI152 brain template.
}
\usage{mni152}
\format{NIfTI-1 format object; voxels: 91x109x91; voxel size: 2x2x2mm}
\references{
  Grabner G., Janke A.L., Budge M.M., Smith D., Pruessner J., and Collins D.L.
  (2006) \emph{Symmetric atlasing and model based segmentation: an application
  to the hippocampus in older adults}. MICCAI, Vol. 4191, pp. 58-66.
}
\keyword{datasets}
