% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/part_coeff.R
\name{part.coeff}
\alias{part.coeff}
\title{Calculate the participation coefficient of each node}
\usage{
part.coeff(g, community.mem)
}
\arguments{
\item{g}{The graph}

\item{community.mem}{The community membership indices of each node}
}
\value{
A vector of the participation coeff's for each node of the graph.
}
\description{
This function calculates the participation coefficient of each node in a
graph. The graph must be split into modules beforehand. The coefficient
should not exceed 1 - (1 / # modules).
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
\references{
Guimera, R. and Amaral, L.A.N. (2005) Cartography of complex
networks: modules and universal roles, Journal of Statistical Mechanics:
Theory and Experiment, 02, P02001.
}

