% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permute_group.R
\name{permute.group}
\alias{permute.group}
\title{Permutation test for group difference of graph measures}
\usage{
permute.group(densities, resids, num.subjs, num.perms = 1000,
  level = c("graph", "vertex"))
}
\arguments{
\item{densities}{A vector of graph densities to loop through}

\item{resids}{A data.table of the residuals (from \code{\link{get.resid}})}

\item{num.subjs}{A vector of length 2 indicating group sizes}

\item{num.perms}{The number of permutations to perform (default: 1e3)}

\item{level}{A character string for the attribute level to calculate
differences; either 'graph' or 'vertex'}
}
\value{
A data table with values for group differences in modularity, global
efficiency, clustering, average path length, and assortativity
}
\description{
This function draws permutations from linear model residuals to determine the
significance of between-group differences of a global graph measure. This
function is intended for cortical thickness networks (in which there is only
one graph per group).
}
\examples{
\dontrun{
out <- permute.group(densities, m$resids, summary(covars$Group), 1e3, 'graph')
}
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}

