% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permute_vertex.R
\name{permute.vertex}
\alias{permute.vertex}
\title{Permutation test for group difference of graph vertex measures}
\usage{
permute.vertex(g1, g2, alpha = 0.05, N = 1000, measure)
}
\arguments{
\item{g1}{A list of igraph graph objects for group 1}

\item{g2}{A list of igraph graph objects for group 2}

\item{alpha}{The significance level (default: 0.05)}

\item{N}{The number of permutations (default: 1e3)}

\item{measure}{A character string of the measure to test}
}
\value{
A list with the following elements:
\item{g}{A graph representing group differences, with \emph{p.perm} included
as a vertex-level attribute (it is actually 1 - \emph{p.perm})}
\item{p.max}{The proportion of the number of times the maximum t-statistic of
the permuted groups was greater than the observed maximum t-statistic}
\item{thresh}{The \eqn{1 - \alpha} \%ile maximum t-statistic of all permuted values}
}
\description{
This function performs a permutation test looking at the max t-statistic for
a given vertex-level graph attribute (e.g. degree). This is the same
principle as that of Nichols & Holmes (2001) used in voxelwise MRI analysis.
}
\examples{
\dontrun{
g.diff <- permute.groups(g.wt[[1]][[3]], g.wt[[2]][[3]], measure='degree')
}
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
\references{
Nichols TE & Holmes AP (2001). \emph{Nonparametric permutation
tests for functional neuroimaging: A primer with examples.} Human Brain
Mapping, 15(1):1-25.
}
\seealso{
\code{\link{sample}, \link{group.graph.diffs}}
}

