% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brainGraph.R
\name{plot.brainGraph_mtpc}
\alias{plot.brainGraph_mtpc}
\title{Plot a graph with results from MTPC}
\usage{
\method{plot}{brainGraph_mtpc}(x, subgraph = "sig == 1",
  main = paste0(x$outcome, ": ", x$name), subtitle = NULL,
  cex.main = 2, ...)
}
\arguments{
\item{x}{A \code{brainGraph_mtpc} graph object}

\item{subgraph}{Character string specifying a logical condition for vertices
to plot. Default: \code{NULL}}

\item{main}{Character string; the main title. Default: \code{NULL}}

\item{subtitle}{Character string; the subtitle. Default: \code{'default'}}

\item{cex.main}{Numeric indicating the scaling for plot title size (see
\code{\link[graphics]{par}}.}

\item{...}{Other parameters (passed to \code{\link[igraph]{plot.igraph}}).
See \code{\link[igraph]{plot.common}} for details.}
}
\description{
This is a convenience function for plotting a graph based on results from
\code{\link{mtpc}}. There are a few argument defaults: to plot only those
vertices for which \eqn{A_{mtpc} > A_{crit}}; a plot title with the outcome
measure and contrast name, and to omit the plot subtitle.
}
\seealso{
Other Plotting functions: \code{\link{plot.brainGraphList}},
  \code{\link{plot.brainGraph_GLM}},
  \code{\link{plot.brainGraph_NBS}},
  \code{\link{plot.brainGraph_mediate}},
  \code{\link{plot.brainGraph}},
  \code{\link{plot_brainGraph_gui}},
  \code{\link{plot_brainGraph_multi}}
}
\concept{Plotting functions}
