% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{plot.corr_mats}
\alias{plot.corr_mats}
\title{Plot raw or thresholded correlation matrices}
\usage{
\method{plot}{corr_mats}(x, type = c("thresholded", "raw"),
  thresh.num = 1L, ordered = TRUE, order.by = "lobe",
  graphs = NULL, grp.names = NULL, legend.title = "", ...)
}
\arguments{
\item{x}{A \code{corr_mats} object}

\item{type}{Character string indicating whether to plot raw or thresholded
(binarized) matrices. Default: \code{'raw'}}

\item{thresh.num}{Integer specifying which threshold to plot (if
\code{type='thresholded'}). Default: \code{1L}}

\item{ordered}{Logical indicating whether to order the vertices by some
grouping. Default: \code{TRUE}}

\item{order.by}{Character string indicating how to group vertices. Default:
\code{'lobe'} (i.e., group into lobes)}

\item{graphs}{Integer vector specifying the ordering. Default: \code{NULL}}

\item{grp.names}{Character vector specifying the names of each group of
vertices. Default: \code{NULL}}

\item{legend.title}{Character string for the legend title. Default is to
leave blank}

\item{...}{Unused}
}
\description{
Plot raw or thresholded correlation matrices
}
