% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_brainGraph_axial.R
\name{plot_brainGraph_axial}
\alias{plot_brainGraph_axial}
\title{Draw axial slice of the MNI152 T1 image.}
\usage{
plot_brainGraph_axial(save = 0, fname = NULL, z = 46)
}
\arguments{
\item{save}{Binary indicating whether or not a png file should be saved}

\item{fname}{The name of the file to be saved}

\item{z}{The z-coordinate of the slice to use (defaults to 46, the center)}
}
\description{
This function draws an axial slice from the MNI152 T1 image, to plot the
nodes of a graph over it. It will optionally write to a filename for output.
}
\seealso{
\code{\link{plot_brainGraph_sagittal}}
}

