% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brainGraph_gui.R
\name{plot_brainGraph_gui}
\alias{plot_brainGraph_gui}
\title{GUI for plotting graphs overlaid on an MNI152 image or in a circle}
\usage{
plot_brainGraph_gui()
}
\description{
This function creates a GUI for plotting graphs over an image from the MNI152
template. It gives the user control over several plotting parameters. Also
possible is a circular plot (in addition to the axial and sagittal views). It
is necessary for the graphs to have an \emph{atlas} attribute, and several
vertex- and edge-level attributes (set by
\code{\link{set_brainGraph_attr}}).
}
\seealso{
Other Plotting functions: \code{\link{Plotting GLM graphs}},
  \code{\link{plot.brainGraphList}},
  \code{\link{plot.brainGraph}},
  \code{\link{plot_brainGraph_multi}}
}
\concept{Plotting functions}
