% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_community.R
\name{plot_community}
\alias{plot_community}
\title{Plot a given community}
\usage{
plot_community(g, n, ...)
}
\arguments{
\item{g}{The graph}

\item{n}{The community to be plotted (integer or vector)}

\item{...}{Other parameters (passed to \code{\link{plot_brainGraph}})}
}
\description{
This function plots a single (or multiple) community, if you want to see e.g.
the first community and nothing else, or the first two, etc.
}

