% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brainGraph.R
\name{plot_mni}
\alias{plot_mni}
\title{Plot a slice of the MNI152 template}
\usage{
plot_mni(plane, hemi)
}
\arguments{
\item{plane}{Character string indicating which orientation to plot.
Default: \code{'axial'}}

\item{hemi}{Character string indicating which hemisphere to plot. Default:
\code{'both'}}
}
\description{
Helper function that plots a single slice from the MNI152 template.
}
\keyword{internal}
