% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_neighborhood.R
\name{plot_neighborhood}
\alias{plot_neighborhood}
\title{Plot the neighborhood of a given vertex.}
\usage{
plot_neighborhood(g, n, ...)
}
\arguments{
\item{g}{The graph to be plotted}

\item{n}{Index (integer) of the vertex whose neighborhood is to be plotted}

\item{...}{Other parameters (passed to \code{\link{plot.adj}})}
}
\description{
This function plots only the neighborhood of a single given vertex, e.g. if
you want to see the neighborhood of the left precuneus and nothing else.
}

