% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_thickness.R
\name{plot_thickness}
\alias{plot_thickness}
\title{Plot group distributions of cortical thickness for a given brain region}
\usage{
plot_thickness(dat, cur.region, type = c("histogram", "violin"))
}
\arguments{
\item{dat}{A data table of cortical thicknesses; needs columns for 'Group',
'region', and 'thickness'}

\item{cur.region}{A character string or integer of the brain region to plot;
if integer, the region is chosen from the data table based on the index}

\item{type}{A character string indicating the plot type; either 'histogram'
or 'violin'}
}
\value{
A ggplot object
}
\description{
This function takes a "tidied" dataset of cortical thickness and plots a
histogram or violin plot for 1 or more groups.
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
\seealso{
\code{\link[ggplot2]{geom_histogram}, \link[ggplot2]{geom_vline}}
}

