% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_global.R
\name{rename_levels}
\alias{rename_levels}
\title{Rename the levels of global metrics in a data.table}
\usage{
rename_levels(DT)
}
\arguments{
\item{DT}{A \code{data.table}}
}
\value{
The levels of the \code{variable} column after being changed
}
\description{
Rename the levels of global metrics in a data.table
}
\keyword{internal}
