% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation}
\alias{rotation}
\title{Apply a rotation matrix to a set of points}
\usage{
rotation(x, theta)
}
\arguments{
\item{x}{A matrix with 2 columns of the points to rotate}

\item{theta}{The angle to apply}
}
\value{
A matrix with 2 columns of the points' new locations
}
\description{
This function takes a set of points and applies a rotation matrix (e.g. will
rotate points 90 deg. if given "pi/2" as input)
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}

