% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resid.R
\name{rstudent_mat}
\alias{rstudent_mat}
\title{Calculate studentized residuals with matrix input}
\usage{
rstudent_mat(X, y, lev, n, p)
}
\arguments{
\item{X}{Numeric matrix; the design matrix}

\item{y}{Numeric vector; the outcome variable}

\item{lev}{Numeric vector; the leverage of the "hat" matrix}

\item{n}{Integer; number of rows in \code{X}}

\item{p}{Integer; number of columns in \code{X}}
}
\description{
Calculate studentized residuals with matrix input
}
\keyword{internal}
