% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set.brainGraph.attributes}
\alias{set.brainGraph.attributes}
\title{Set a number of graph and vertex attributes useful in MRI analyses}
\usage{
set.brainGraph.attributes(g, atlas = NULL, coords = NULL, rand = FALSE)
}
\arguments{
\item{g}{A graph}

\item{atlas}{A character vector indicating which atlas was used for the nodes}

\item{coords}{A matrix of the spatial coordinates}

\item{rand}{Logical indicating whether the graph is random or not}
}
\value{
g A copy of the same graph, with the following attributes:
\item{Graph-level}{Atlas, density, connected component sizes, diameter,
number of triangles, transitivity, average path length, assortativity,
global efficiency, modularity, and rich-club coefficient}
\item{Vertex-level}{Degree, betweenness centrality, eigenvector centrality,
transitivity (local), coreness, local efficiency, color, community membership,
participation coefficient, within-module degree z-score, and coordinates (x,
y, and z)}
\item{Edge-level}{Color, lobe color}
}
\description{
This function will set a number of graph and vertex attributes of a given
igraph object.
}
\seealso{
\code{\link{clusters}, \link{graph.motifs}, \link{diameter},
\link{centralization.betweenness}, \link{centralization.evcent},
\link{subgraph.centrality}, \link{hub.score}, \link{authority.score},
\link{transitivity}, \link{average.path.length}, \link{assortativity.degree},
\link{global.eff}, \link{local.eff}, \link{rich.club.coeff},
\link{edge.betweenness.community}, \link{color.edges}, \link{part.coeff},
\link{within.module.deg.z.score},\link{graph.coreness}}
}

