% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainGraph_GLM.R
\name{setup_randomise}
\alias{setup_randomise}
\title{Helper function to setup for randomise}
\usage{
setup_randomise(X, con.mat, nC)
}
\arguments{
\item{X}{Numeric matrix, if you wish to supply your own design matrix
(default: \code{NULL})}

\item{con.mat}{Numeric matrix specifying the contrast(s) of interest; if
only one contrast is desired, you can supply a vector}

\item{nC}{Integer; the number of contrasts}
}
\description{
\code{setup_randomise} is used to setup the data/objects for any function
that does permutations for GLM-based analysis.
}
\details{
The tasks performed by this function are: separate the design matrix into the
independent variables of interest and nuisance variables (based on the
contrast(s)), and calculate the new contrast(s) based on this new design
matrix.
}
\keyword{internal}
