% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_rand_graph_clust.R
\name{sim.rand.graph.clust}
\alias{sim.rand.graph.clust}
\title{Simulate a random graph with given degree sequence and clustering.}
\usage{
sim.rand.graph.clust(g, cl = g$transitivity, max.iters = 100)
}
\arguments{
\item{g}{An \code{igraph} graph object from which null graphs are simulated}

\item{cl}{The clustering measure (default: transitivity)}

\item{max.iters}{The maximum number of iterations to perform (default: 100)}
}
\value{
An \code{igraph} graph object
}
\description{
This function will simulate a random graph with a given degree sequence and
clustering coefficient. This function calls \link{choose.edges} to decide
which edges will be re-wired.
}
\references{
Bansal S., Khandelwal S., Meyers L.A. (2009) \emph{Exploring
biological network structure with clustered random networks}. BMC
Bioinformatics, 10:405-421.
}
\seealso{
\code{\link[igraph]{rewire}, \link[igraph]{transitivity},
  \link[igraph]{keeping_degseq}}

Other Null graph functions: \code{\link{analysis_random_graphs}},
  \code{\link{choose.edges}}, \code{\link{rich_club_norm}},
  \code{\link{sim.rand.graph.par}}
}
