% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_string}
\alias{split_string}
\title{Add newlines to a character string for printing}
\usage{
split_string(x, max_len = 80L, delim = "+", sep = "\\n")
}
\arguments{
\item{x}{A character string with length 1}

\item{delim}{Character specifying where to end a line if it is longer than
\code{max.len}. Default: \code{'+'}}

\item{sep}{Character specifying what to split by. Default: \code{'\n'} (a
newline character)}

\item{max.len}{Integer; the max length of one line. Default: 80}
}
\description{
The \code{delim} argument determines \emph{where} to insert the separator,
which by default is a newline character.
}
\keyword{internal}
