% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_graph}
\alias{subset_graph}
\title{Subset graphs based on a given logical condition}
\usage{
subset_graph(g, subgraph)
}
\arguments{
\item{g}{A graph object}

\item{subgraph}{Character string specifying an equation for which vertices to
keep}
}
\value{
A graph object
}
\description{
\code{subset_graphs} will subset a given graph based on the given logical
condition(s). This can be a "simple" logical equation, or can include
combinations of \emph{AND} and \emph{OR} (i.e., \code{&} and \code{|}).
}
\keyword{internal}
