% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainGraph_mediate.R
\name{summary.bg_mediate}
\alias{summary.bg_mediate}
\title{Print a summary from a brainGraph mediation analysis}
\usage{
\method{summary}{bg_mediate}(object, mediate = FALSE, region = NULL,
  digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{object}{A \code{bg_mediate} object}

\item{mediate}{Logical indicating whether or not to use the \code{summary}
method from \code{\link[mediation]{mediate}} (default: \code{FALSE}). If
\code{TRUE}, only a single region can be printed.}

\item{region}{Character string specifying which region's results to
summarize; only relevant if \code{level='vertex'} (default: \code{NULL})}

\item{digits}{Integer specifying the number of digits to display for p-values}

\item{...}{Unused}
}
\description{
Print a summary from a brainGraph mediation analysis
}
