% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphs.R
\name{summary.brainGraph}
\alias{summary.brainGraph}
\title{Print a summary of a brainGraph object}
\usage{
\method{summary}{brainGraph}(object, print.attrs = c("all", "none"), ...)
}
\arguments{
\item{object}{A \code{brainGraph} object}

\item{print.attrs}{Character string indicating whether or not to list the
object's attributes (default: \code{all})}

\item{...}{Unused}
}
\description{
Print a summary of a brainGraph object
}
