% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_global.R
\name{summary.brainGraph_boot}
\alias{summary.brainGraph_boot}
\title{Print a summary from a bootstrap analysis}
\usage{
\method{summary}{brainGraph_boot}(object, ...)
}
\arguments{
\item{object}{A \code{brainGraph_boot} object (from
\code{\link{brainGraph_boot}})}

\item{...}{Unused}
}
\description{
Print a summary from a bootstrap analysis
}
