% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_group.R
\name{summary.brainGraph_permute}
\alias{summary.brainGraph_permute}
\title{Print a summary from a permutation analysis}
\usage{
\method{summary}{brainGraph_permute}(object, measure = NULL,
  alternative = c("two.sided", "less", "greater"), alpha = 0.05,
  p.sig = c("p", "p.fdr"), ...)
}
\arguments{
\item{object}{A \code{brainGraph_permute} object (output by
\code{\link{brainGraph_permute}}).}

\item{measure}{Character string of the measure to focus on (if
\code{level='graph'})}

\item{alternative}{Character string, whether to do a two- or one-sided test
(default: \code{'two.sided'})}

\item{alpha}{Numeric; the significance level (default: 0.05)}

\item{p.sig}{Character string specifying which p-value to use for displaying
significant results (default: \code{p})}

\item{...}{Unused}
}
\description{
Print a summary from a permutation analysis
}
