% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpc.R
\name{summary.mtpc}
\alias{summary.mtpc}
\title{Print a summary of MTPC results}
\usage{
\method{summary}{mtpc}(object, contrast = NULL, digits = max(3L,
  getOption("digits") - 2L), print.head = TRUE, ...)
}
\arguments{
\item{object}{A \code{mtpc} object}

\item{contrast}{Integer specifying the contrast to summarize; defaults to
showing results for all contrasts}

\item{digits}{Integer specifying the number of digits to display for p-values}

\item{print.head}{Logical indicating whether or not to print only the first
and last 5 rows of the statistics tables (default: \code{TRUE})}

\item{...}{Unused}
}
\description{
Print a summary of MTPC results
}
