% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBS.R
\name{summary.NBS}
\alias{summary.NBS}
\title{Print a summary of NBS analysis}
\usage{
\method{summary}{NBS}(object, contrast = NULL, digits = max(3L,
  getOption("digits") - 2L), ...)
}
\arguments{
\item{object}{A \code{NBS} object}

\item{contrast}{Integer specifying the contrast to summarize; defaults to
showing results for all contrasts}

\item{digits}{Integer specifying the number of digits to display for p-values}

\item{...}{Unused}
}
\description{
Print a summary of NBS analysis
}
