% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symm_mean}
\alias{symm_mean}
\title{Symmetrize a matrix with the mean of off-diagonal elements}
\usage{
symm_mean(A)
}
\arguments{
\item{A}{Numeric matrix}
}
\value{
Numeric matrix
}
\description{
\code{symm_mean} returns a symmetric matrix in which the off-diagonal
elements \eqn{A[i, j]} and \eqn{A[j, i]} are equal to the mean of the values
in the input matrix.
}
\keyword{internal}
