% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mats.R
\name{symmetrize_mats}
\alias{symmetrize_mats}
\title{Create a symmetric matrix}
\usage{
symmetrize_mats(A, symm.by = c("max", "min", "avg"))
}
\arguments{
\item{A}{Numeric matrix}

\item{symm.by}{Character string; how to create symmetric off-diagonal
elements (default: \code{max})}
}
\description{
\code{symmetrize_mats} will symmetrize a numeric matrix by assigning the
off-diagonal elements values of either the \code{max}, \code{min}, or
\code{average} of \eqn{\{A(i, j), A(j, i)\}}. The default is \code{max}
because that is the default for
\code{\link[igraph]{graph_from_adjacency_matrix}}.
}
\seealso{
\code{\link[igraph]{graph_from_adjacency_matrix}}

Other Matrix functions: \code{\link{cor.diff.test}},
  \code{\link{create_mats}}
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
