% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/targeted_attack.R
\name{targeted.attack}
\alias{targeted.attack}
\title{Perform a targeted attack of a graph}
\usage{
targeted.attack(g, type = c("vertex", "edge"), measure = c("btwn.cent",
  "degree"))
}
\arguments{
\item{g}{The igraph graph object of interest}

\item{type}{A character string; either 'vertex' or 'edge' removals}

\item{measure}{A character string; sort by either 'btwn.cent' or 'degree'}
}
\value{
A vector representing the percent of maximal component size compared
to the graph's original maximal component
}
\description{
This function performs a "targeted attack" of a graph. It will sort the
vertices by either degree or betweenness centrality, and successively remove
the top vertices. There is also an option to remove edges successively in
terms of edge betweenness. Then it calculates the size of the largest
component.
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}

