% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/update_adj.R
\name{update.adj}
\alias{update.adj}
\title{Function to dynamically plot a graph}
\usage{
\method{update}{adj}(graphname1, graphname2, vertLabels, vertSize, edgeWidth,
  edgeDiffs, vertColor, firstplot, secondplot, vertSize.const = NULL,
  plotFunc, comm = NULL, comboNeighb, hemi, lobe, orient1, orient2,
  showDiameter, slider1, slider2)
}
\description{
This function is called by \link{plot.adj.gui} to update a plot on-the-fly.
It updates both of the plots by calling the helper function "make.plot".
}

