% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{update_fs_names}
\alias{update_fs_names}
\title{Update column names in a Freesurfer table}
\usage{
update_fs_names(filename, modality, parcellation, hemi, exclude.subs)
}
\description{
Updates column names by abbreviating to match with the names in the
\code{brainGraph} atlases.
}
\keyword{internal}
