% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_stats.R
\name{vif.bg_GLM}
\alias{vif.bg_GLM}
\title{Variance inflation factors for \code{bg_GLM} objects}
\usage{
vif.bg_GLM(mod, ...)
}
\arguments{
\item{mod}{A \code{bg_GLM} object}

\item{...}{Unused}
}
\value{
A named array of VIFs; names of the 3rd dimension are regions
}
\description{
Variance inflation factors for \code{bg_GLM} objects
}
