% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vulnerability.R
\name{vulnerability}
\alias{vulnerability}
\title{Calculate the vulnerability of each vertex of a graph}
\usage{
vulnerability(g)
}
\arguments{
\item{g}{The igraph graph object of interest}
}
\value{
A vector of length equal to the number of vertices in \emph{g}
}
\description{
This function calculates the vulnerability of the vertices of a graph. Here,
vulnerability is considered to be the proportional drop in global efficiency
when a given node is removed from the graph. The vulnerability of the graph
is considered the maximum across all vertices.
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
\references{
Latora V., Marchiori M. (2005) \emph{Variability and protection
of infrastructure networks}. Physical Review E, 71:015103.
}
\seealso{
\code{\link{graph.efficiency}}
}

