% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write.brainnet}
\alias{write.brainnet}
\title{Write files to be used for visualization with BrainNet Viewer}
\usage{
write.brainnet(g, node.color = c("none", "community"), node.size = "degree")
}
\arguments{
\item{g}{A graph}

\item{node.size}{Character string indicating what size the nodes should be;
can be any vertex-level attribute (see \link{list.vertex.attributes})}

\item{node.color}{Character string indicating whether to color the nodes or
not; can be "none" or "community"}
}
\description{
This function will write the "node" and "edge" files necessary for
visualization with the BrainNet Viewer software.
}

