% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_brainnet.R
\name{write_brainnet}
\alias{write_brainnet}
\title{Write files to be used for visualization with BrainNet Viewer}
\usage{
write_brainnet(g, node.color = "none", node.size = "constant",
  edge.wt = NULL, file.prefix = "")
}
\arguments{
\item{g}{The \code{igraph} graph object of interest}

\item{node.color}{Character string indicating whether to color the vertices or
not (default: \code{'none'})}

\item{node.size}{Character string indicating what size the vertices should be;
can be any vertex-level attribute (default: \code{'constant'})}

\item{edge.wt}{Character string indicating the edge attribute to use to
return a weighted adjacency matrix (default: \code{NULL})}

\item{file.prefix}{Character string for the basename of the \emph{.node} and
\emph{.edge} files that are written}
}
\description{
This function will write the \emph{.node} and \emph{.edge} files necessary
for visualization with the BrainNet Viewer software (see Reference below).
}
\details{
For the \emph{.node} file, there are 6 columns:
\itemize{
\item \emph{Column 1}: x-coordinates
\item \emph{Column 2}: y-coordinates
\item \emph{Column 3}: z-coordinates
\item \emph{Column 4}: Vertex color
\item \emph{Column 5}: Vertex size
\item \emph{Column 6}: Vertex label
}
The \emph{.edge} file is the graph's associated adjacency matrix; a weighted
adjacency matrix can be returned by using the \code{edge.wt} argument.
}
\examples{
\dontrun{
write_brainnet(g, node.color='community', node.size='degree',
  edge.wt='t.stat')
}
}
\references{
Xia M, Wang J, He Y (2013). \emph{BrainNet Viewer: a network
visualization tool for human brain connectomics}. PLoS One, 8(7):e68910.
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
