% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_diff.R
\name{auc_diff}
\alias{auc_diff}
\title{Difference in the area-under-the-curve of two vectors}
\usage{
auc_diff(x, y)
}
\arguments{
\item{x}{Numeric vector of the x-values}

\item{y}{A 2-column numeric matrix; each column contains the values for one
group}
}
\value{
A numeric value of the difference between two groups
}
\description{
This function takes two vectors, calculates the area-under-the-curve (AUC),
and calculates the difference between the two.
}

