\name{brainsuite}
\docType{data}
\alias{brainsuite}
\title{Coordinates for data from the BrainSuite software}
\description{
  This is a list of spatial coordinates for the BrainSuite software, along with
  indices for the major lobes of the brain.
}
\usage{brainsuite}
\format{List, 17 elements}
\references{
  Shattuck DW and Leahy RM (2002) \emph{BrainSuite: an automated cortical
  surface identification tool}. Medical Image Analysis, 8(2):129-142.
  Pantazis D, Joshi AA, Jintao J, Shattuck DW, Bernstein LE, Damasio H, and
  Leahy RM. (2009) \emph{Comparison of landmark-based and automatic methods
      for cortical surface registration}. NeuroImage, 49(3):2479-2493.
}
\keyword{datasets}
