% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_resid.R
\name{check.resid}
\alias{check.resid}
\title{Check model residuals for each brain region.}
\usage{
check.resid(resids)
}
\arguments{
\item{resids}{Data table of model residuals for all brain regions}
}
\description{
This function checks the model residuals for each brain region in the
analysis. It simply does a qqplot of the studentized residuals.
}
\seealso{
\code{\link[stats]{qqnorm}}
}

