% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coeff_var}
\alias{coeff_var}
\title{Calculate coefficient of variation}
\usage{
coeff_var(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Calculates the \emph{coefficient of variation}, defined as
\deqn{CV(x) = \frac{sd(x)}{mean(x)}}
}
