% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/color_edges.R
\name{color.edges}
\alias{color.edges}
\title{Color graph edges}
\usage{
color.edges(g, memb)
}
\arguments{
\item{g}{The graph to get its edges colored}

\item{memb}{An integer vector indicating vertex group membership}
}
\value{
A character vector of colors for each edge in the graph
}
\description{
This function takes the community membership of a given graph, and assigns
to the edges a specific color (the same as the vertex membership colors).
Edges that connect vertices of two different groups are colored gray. Also
works for the major lobes of the brain, plus insula, subcortical gray matter,
cingulate, limbic lobe (if included in the specific brain atlas).
}

