% Generated by roxygen2 (4.0.1): do not edit by hand
\name{color.edges}
\alias{color.edges}
\title{Color the edges of a graph based on community membership or major lobe.}
\usage{
color.edges(g, comm, lobes = NULL, lobe.cols = NULL)
}
\arguments{
\item{g}{The graph to get its edges colored}

\item{comm}{An integer vector indicating community membership}

\item{lobes}{An integer vector of the lobe membership for each vertex}

\item{lobe.cols}{A character vector of the colors each lobe should take}
}
\value{
A character vector of colors for each edge in the graph
}
\description{
This function takes the community membership of a given graph, and assigns
to the edges a specific color (the same as the vertex membership colors).
Currently only works for up to 9 communities (colored: red, green, blue,
yellow, magenta, orangered, lightgreen, lightblue, lightyellow). Edges that
connect two different communities are colored gray. Also works for the major
lobes of the brain (plus insula).
}

