% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color_vertices.R
\name{color.vertices}
\alias{color.vertices}
\title{Color graph vertices}
\usage{
color.vertices(memb, atlas = NULL)
}
\arguments{
\item{memb}{An integer vector representing membership of e.g. a community}

\item{atlas}{A character string}
}
\value{
A character vector with length equal to the number of communities,
lobes, components, etc.
}
\description{
This function takes an integer vector (representing membership of a community
or component), or a character string indicating the brain atlas used (e.g. for
the major lobes of the brain), and creates a character vector of colors for
each community/module, component, etc.
}

