% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vertices.R
\name{color.vertices}
\alias{color.vertices}
\title{Color graph vertices}
\usage{
color.vertices(memb)
}
\arguments{
\item{memb}{An integer vector representing membership of e.g. a community}
}
\value{
A character vector with length equal to the number of communities,
lobes, components, etc.
}
\description{
This function takes an integer vector (representing membership of a community
or component) and creates a character vector of colors for each
community/module, component, etc. This only assigns a color to groups with at
least 2 members; isolated vertices will be colored 'gray'.
}

