% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/color_vertices.R
\name{color.vertices}
\alias{color.vertices}
\title{Color graph vertices}
\usage{
color.vertices(memb, cols)
}
\arguments{
\item{memb}{An integer vector representing membership of e.g. a community}

\item{cols}{A character vector of the colors each vertex group should take}
}
\value{
A character vector with length equal to the number of communities,
lobes, components, etc.
}
\description{
This function takes an integer vector (representing membership of a community
or component) and creates a character vector of colors for each
community/module, component, etc.
}

