% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract_brainGraph.R
\name{contract_brainGraph}
\alias{contract_brainGraph}
\title{Contract graph vertices based on brain lobe and hemisphere}
\usage{
contract_brainGraph(g)
}
\arguments{
\item{g}{An \code{igraph} graph object}
}
\value{
A new \code{igraph} graph object
}
\description{
This function creates a new graph after merging multiple vertices based on
brain lobe and hemisphere membership. The new vertex size is equal to the
number of vertices in each lobe. The x- and y- coordinates of the new
vertices are equal to the mean of the lobe vertices of the original graph.
The new edge weight is equal to the number of inter-lobular connections of
the original graph.
}
\seealso{
\code{\link[igraph]{contract}}
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
