% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_resid.R
\name{get.resid}
\alias{get.resid}
\title{Run linear models of thickness data to get model residuals.}
\usage{
get.resid(thicknesses, covars)
}
\arguments{
\item{thicknesses}{Data frame of all thickness data. First column must be
subject ID (or something similar)}

\item{covars}{Data frame of covariates for the linear model}
}
\value{
A list with components:
\item{models}{The \link{lm} objects for each brain region.}
\item{resids}{Data table of residuals for all subjects and brain regions.}
}
\description{
This function runs linear models on the thickness data for each region, in
order to adjust for relevant variables (e.g. age, sex, group, etc.).
}

