% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata}
\alias{get_metadata}
\title{Add metadata to a list-like object}
\usage{
get_metadata(object)
}
\arguments{
\item{object}{A list-like object}
}
\value{
The same object with version, system, and date information added
}
\description{
\code{get_metadata} adds metadata to a list-like object.
}
\details{
The information added are:
\itemize{
  \item{version}{R, brainGraph, and igraph versions}
  \item{sys}{System information}
  \item{date}{The date and time of creation}
}
}
\keyword{internal}
