\name{global.eff}
\alias{global.eff}
\title{Calculate the global efficiency of a graph.}
\usage{
global.eff(g)
}
\arguments{
  \item{g}{The adjacency graph on which to calculate
  efficiency}
}
\value{
A number; the global efficiency of the graph.
}
\description{
This function calculates the global efficiency of a graph.
}

