\name{local.eff}
\alias{local.eff}
\title{Calculate the local efficiency of a graph.}
\usage{
local.eff(g)
}
\arguments{
  \item{g}{the adjacency graph on which to calculate
  efficiency}
}
\value{
A vector of the local efficiencies for each node of the
graph.
}
\description{
This function calculates the local efficiency of each node
of a graph.
}

