% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_empty_brainGraph.R
\name{make_empty_brainGraph}
\alias{make_empty_brainGraph}
\title{Create an empty graph with attributes for brainGraph}
\usage{
make_empty_brainGraph(g)
}
\arguments{
\item{g}{An \code{igraph} graph object}
}
\value{
An empty \code{igraph} graph object with several additional
  attributes
}
\description{
This function creates an empty graph and includes some graph-, vertex-, and
edge-level attributes that are important for \code{brainGraph} functions.
Basically a wrapper for \code{\link[igraph]{make_empty_graph}}.
}
\details{
The input graph must have the graph attribute \emph{atlas} and vertex
attribute \emph{name} already.
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
\seealso{
\code{\link[igraph]{make_empty_graph}}
}

