% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Methods}
\alias{Methods}
\alias{groups.brainGraphList}
\alias{groups.brainGraph_resids}
\alias{groups.corr_mats}
\alias{nobs.brainGraphList}
\alias{nobs.brainGraph_resids}
\alias{case.names.brainGraph_resids}
\alias{region.names}
\alias{region.names.data.table}
\alias{nregions}
\title{Return group names}
\usage{
\method{groups}{brainGraphList}(x)

\method{groups}{brainGraph_resids}(x)

\method{groups}{corr_mats}(x)

\method{nobs}{brainGraphList}(object, ...)

\method{nobs}{brainGraph_resids}(object, ...)

\method{case.names}{brainGraph_resids}(object, ...)

region.names(object)

\method{region.names}{data.table}(object)

nregions(object)
}
\arguments{
\item{x, object}{An object}

\item{...}{Unused}
}
\description{
\code{groups} returns the \dQuote{Group} graph attribute for each graph or
observation in the object.

\code{region.names} is a generic method for extracting region names from
various \code{brainGraph} objects. These are generally convenience functions.

\code{nregions} is a generic method for extracting the number of regions from
various \code{brainGraph} objects.
}
\details{
For a \code{data.table}, \code{region.names} assumes that it contains a
\emph{factor} column named \code{region}.
}
