% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainGraph_GLM.R
\name{partition}
\alias{partition}
\title{Partition a design matrix into columns of interest and nuisance}
\usage{
partition(M, con.mat, method = c("beckmann", "guttman"))
}
\arguments{
\item{M}{Numeric matrix; the full design matrix}

\item{con.mat}{Numeric matrix; the contrast matrix}

\item{method}{Character string; the method of partitioning (default:
\code{beckmann})}
}
\description{
Partition a design matrix into columns of interest and nuisance
}
\references{
Guttman I. Linear Models: An Introduction. Wiley, New York, 1982.

Smith SM, Jenkinson M, Beckmann C, Miller K, Woolrich M (2007).
  \emph{Meaningful design and contrast estimability in fMRI.} NeuroImage,
  34(1):127-36.
}
\keyword{internal}
