% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_adj_gui.R
\name{plot.adj.gui}
\alias{plot.adj.gui}
\title{GUI for plotting graphs overlaid on an MNI152 image or in a circle.}
\usage{
\method{plot}{adj.gui}()
}
\description{
This function creates a GUI for plotting graphs over an image from the MNI152
template. It gives the user control over several plotting parameters. Also
possible is a circular plot (in addition to the axial and sagittal views).
}

