% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.adj}
\alias{plot.adj}
\title{Plots an adjacency matrix with a specific spatial layout.}
\usage{
\method{plot}{adj}(g, vertex.size = 10, edge.width = 2, rescale = F,
  ylim = c(-1.5, 1.5), ...)
}
\arguments{
\item{g}{The graph to plot}

\item{vertex.size}{The size of the vertices}

\item{edge.width}{The size of the edges}

\item{rescale}{A logical constant, whether to rescale the coordinates}

\item{ylim}{A vector giving limits for the vertical axis}

\item{...}{Other parameters (passed to \code{\link{plot}}).}
}
\description{
This function plots an adjacency matrix when the spatial layout of the nodes
is important. It is really just a wrapper for \code{\link{plot}}, but it uses
as a default spatial layout coordinates that work for the brain. This means
that \code{\link{set.brainGraph.attributes}} needs to be run on the graph,
and a valid set of coordinates provided for the vertices. Most of the
parameters here are for \code{\link{plot.igraph}}.
}

