\name{plot.neighborhood}
\alias{plot.neighborhood}
\title{Plot the neighborhood of a given vertex.}
\usage{
\method{plot}{neighborhood}(g, v)
}
\arguments{
  \item{g}{The total adjacency matrix}

  \item{v}{Character string of the vertex name whose
  neighborhood is to be plotted}
}
\description{
This function plots only the neighborhood of a single given
vertex, if you want to see e.g. the neighborhood of the
left precuneus and nothing else.
}

