% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.neighborhood}
\alias{plot.neighborhood}
\title{Plot the neighborhood of a given vertex.}
\usage{
\method{plot}{neighborhood}(g, v)
}
\arguments{
\item{g}{the total adjacency matrix}

\item{v}{character string of the vertex name whose neighborhood is to be
plotted}
}
\description{
This function plots only the neighborhood of a single given vertex.
}

