% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_brainGraph_mni.R
\name{plot_brainGraph_mni}
\alias{plot_brainGraph_mni}
\title{Draw an axial or sagittal slice of the MNI152 T1 image}
\usage{
plot_brainGraph_mni(plane = c("axial", "sagittal"), slice, hemi = c("L",
  "R"), save = FALSE, fname = NULL)
}
\arguments{
\item{plane}{Character string, either 'axial' or 'sagittal'}

\item{slice}{The x or z-coordinate of the slice to use}

\item{hemi}{Character string, either 'L' or 'R'}

\item{save}{Logical indicating whether or not a png file should be saved
(default: FALSE)}

\item{fname}{The name of the file to be saved}
}
\description{
This function draws an axial or sagittal slice from the MNI152 T1 image, to
plot the vertices of a graph over it. It will optionally write to a filename
for output.
}
\seealso{
\code{\link[oro.nifti]{image.nifti}}
}

