% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_corr_mat.R
\name{plot_corr_mat}
\alias{plot_corr_mat}
\title{Plot correlation matrices}
\usage{
plot_corr_mat(c1, c2, ordered = TRUE, type = c("comm", "lobe"), g1 = NULL,
  g2 = NULL, groups = c("Group 1", "Group 2"))
}
\arguments{
\item{c1}{The correlation matrix for group 1}

\item{c2}{The correlation matrix for group 2}

\item{ordered}{A logical indicating whether or not to order vertices
(default:TRUE)}

\item{type}{Character string, either 'comm' or 'lobe'}

\item{g1}{An igraph graph object; not required if \emph{ordered} is FALSE}

\item{g2}{An igraph graph object; not required if \emph{ordered} is FALSE}

\item{groups}{A character vector of group names}
}
\value{
A list of ggplot objects
}
\description{
This function will plot correlation matrices in the form of a ``heatmap''. It
will work for to two groups, and you have the choice to plot the vertices in
an order based on either community or lobe membership.
}
\examples{
\dontrun{
mat.plots <- plot_corr_mat(corrs.group1[[N]]$r.thresh,
                           corrs.group2[[N]]$r.thresh, g1=g1[[N]],
                           g2=g2[[N]], groups=groups)
}
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
\seealso{
\code{\link[ggplot2]{geom_tile}}
}

