% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_thickness.R
\name{plot_thickness}
\alias{plot_thickness}
\title{Plot group distributions of cortical thickness for a given brain region}
\usage{
plot_thickness(dat, regions, type = c("violin", "histogram"),
  all.vals = TRUE)
}
\arguments{
\item{dat}{A data table of cortical thicknesses; needs columns for 'Group',
'region', and 'thickness'}

\item{regions}{A vector of character strings or integers of the brain
region(s) to plot; if integer, the region(s) is/are chosen from the atlas
data table based on the index}

\item{type}{A character string indicating the plot type; either 'histogram'
or 'violin'}

\item{all.vals}{A logical indicating whether or not to plot horizontal lines
for all observations (only valid for 'violin' plots) (default: TRUE)}
}
\value{
A ggplot object
}
\description{
This function takes a "tidied" dataset of cortical thickness and plots a
histogram or violin plot for 1 or more groups.
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
\seealso{
\code{\link[ggplot2]{geom_histogram}, \link[ggplot2]{geom_vline}}
}

