% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainGraph_GLM.R
\name{randomise}
\alias{randomise}
\title{Randomize and fit a model and find the maximum statistic}
\usage{
randomise(ctype, N, perms, DT, nC, measure, X, con.mat, alternative)
}
\arguments{
\item{N}{Integer; number of permutations to create (default: 5e3)}

\item{perms}{Matrix of permutations, if you would like to provide your own
(default: \code{NULL})}

\item{DT}{\code{data.table} with outcome variables}

\item{measure}{Character string of the graph measure of interest}

\item{X}{Numeric matrix, if you wish to supply your own design matrix
(default: \code{NULL})}

\item{con.mat}{Numeric matrix specifying the contrast(s) of interest; if
only one contrast is desired, you can supply a vector}

\item{alternative}{Character string, whether to do a two- or one-sided test
(default: \code{'two.sided'})}
}
\description{
Randomize and fit a model and find the maximum statistic
}
\keyword{internal}
