% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resid.R
\name{rstudent_mat}
\alias{rstudent_mat}
\title{Calculate studentized residuals with matrix input}
\usage{
rstudent_mat(lmvars, y)
}
\arguments{
\item{lmvars}{List containing: \code{X} (design matrix); \code{lev}
(leverage); \code{n} (num. observations); \code{p} (num. parameters)}

\item{y}{Numeric vector; the outcome variable}
}
\description{
Calculate studentized residuals with matrix input
}
\keyword{internal}
