% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_brainGraph_attributes.R
\name{set.brainGraph.attributes}
\alias{set.brainGraph.attributes}
\title{Set a number of graph and vertex attributes useful in MRI analyses}
\usage{
set.brainGraph.attributes(g, atlas = NULL, rand = FALSE)
}
\arguments{
\item{g}{An igraph object}

\item{atlas}{A character vector indicating which atlas was used for the nodes}

\item{rand}{Logical indicating if the graph is random or not (default: FALSE)}
}
\value{
g A copy of the same graph, with the following attributes:
\item{Graph-level}{Package version, atlas, density, connected component sizes,
diameter, # of triangles, transitivity, average path length, assortativity,
clique number, global & local efficiency, modularity, vulnerability, hub score,
rich-club coefficient, and edge asymmetry}
\item{Vertex-level}{Degree, strength, betweenness/eigenvector/subgraph and
leverage centralities, transitivity (local), coreness, local & nodal
efficiency, color (community), color (lobe), color (component), membership
(community), membership (component), participation coefficient, within-module
degree z-score, vulnerability, and coordinates (x, y, and z)}
\item{Edge-level}{Color (community), color (lobe), color (component), edge
betweenness, Euclidean distance (in mm)}
}
\description{
This function will set a number of graph, vertex, and edge attributes of a
given igraph object.
}
\author{
Christopher G. Watson, \email{cgwatson@bu.edu}
}
\seealso{
\code{\link{components}, \link{graph.motifs}, \link{diameter},
\link[igraph]{clique_num}, \link{centralization.betweenness},
\link{edge.betweenness}, \link{centralization.evcent},
\link{subgraph.centrality}, \link{hub.score}, \link{authority.score},
\link{transitivity}, \link{average.path.length}, \link{assortativity.degree},
\link{graph.efficiency}, \link{rich.club.coeff},
\link{edge.betweenness.community}, \link{color.edges}, \link{part.coeff},
\link{within_module_deg_z_score},\link{graph.coreness},\link{spatial.dist},
\link{vulnerability}, \link{centr_lev}, \link{edge_asymmetry},
\link[igraph]{graph.knn}}
}

