% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set_brainGraph_attributes.R
\name{set.brainGraph.attributes}
\alias{set.brainGraph.attributes}
\title{Set a number of graph and vertex attributes useful in MRI analyses}
\usage{
set.brainGraph.attributes(g, atlas = NULL, coords = NULL, rand = FALSE)
}
\arguments{
\item{g}{An igraph object}

\item{atlas}{A character vector indicating which atlas was used for the nodes}

\item{coords}{A matrix of the spatial coordinates}

\item{rand}{Logical indicating whether the graph is random or not}
}
\value{
g A copy of the same graph, with the following attributes:
\item{Graph-level}{Atlas, density, connected component sizes, diameter,
number of triangles, transitivity, average path length, assortativity,
clique sizes, global & local efficiency, modularity, hub score, and rich-club
coefficient}
\item{Vertex-level}{Degree, strength, betweenness centrality, eigenvector
centrality, transitivity (local), coreness, local & nodal efficiency, color
(community), color (lobe), color (component), membership (community),
membership (component), participation coefficient, within-module degree
z-score, and coordinates (x, y, and z)}
\item{Edge-level}{Color (community), color (lobe), color (component), edge
betweenness}
}
\description{
This function will set a number of graph, vertex, and edge attributes of a
given igraph object.
}
\seealso{
\code{\link{components}, \link{graph.motifs}, \link{diameter},
\link{cliques}, \link{centralization.betweenness}, \link{edge.betweenness},
\link{centralization.evcent}, \link{subgraph.centrality}, \link{hub.score},
\link{authority.score}, \link{transitivity}, \link{average.path.length},
\link{assortativity.degree}, \link{graph.efficiency}, \link{rich.club.coeff},
\link{edge.betweenness.community}, \link{color.edges}, \link{part.coeff},
\link{within.module.deg.z.score},\link{graph.coreness}}
}

