% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_rand_graph_clust.R
\name{sim.rand.graph.clust}
\alias{sim.rand.graph.clust}
\title{Simulate a random graph with given degree sequence and clustering.}
\usage{
sim.rand.graph.clust(graph, d, cl)
}
\arguments{
\item{graph}{The graph from which null graphs are simulated}

\item{d}{The degree sequence}

\item{cl}{The clustering measure (clustering coeff, transitivity, etc.)}
}
\value{
A random graph
}
\description{
This function will simulate a random graph with a given degree sequence and
clustering coefficient. This function calls \link{choose.edges} to decide
which edges will be re-wired.
}
\seealso{
\code{\link{degree.sequence.game}, \link{choose.edges},
\link{rewire}, \link{transitivity}, \link{keeping_degseq}}
}

