% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex_attr_dt.R
\name{vertex_attr_dt}
\alias{vertex_attr_dt}
\title{Create a data table with graph vertex measures}
\usage{
vertex_attr_dt(g, group = NULL)
}
\arguments{
\item{g}{An \code{igraph} graph object}

\item{group}{A character string indicating group membership (default: NULL)}
}
\value{
A data table; each row is for a different vertex
}
\description{
This is just a helper function that creates a data table in which each row is
a vertex and each column is a different network measure (degree, centrality,
etc.).
}
\seealso{
\code{\link[igraph]{vertex_attr}, \link[igraph]{vertex_attr_names},
\link[igraph]{as_data_frame}}
}

